lakshya-agent/requirements.txt
google-generativeai[adk]
python-dotenv
google-adk
tools
google-generativeai
adk
pandas
google-adk[memory]
